rename `ValueSubject` to `Variable`
