+++ test/dart_observable/dart_observable_test.dart
@@ -24,7 +24,7 @@ import 'observers/observer_default_constructor_test.dart' as observer_default_co
 import 'subjects/publisher_test.dart' as publisher_test;
 import 'subjects/replayer_test.dart' as replayer_test;
 import 'subjects/subject_test.dart' as subject_test;
-import 'subjects/value_subject_test.dart' as value_subject_test;
+import 'subjects/variable_test.dart' as variable_test;
 import 'models/async_test.dart' as async_test;
 
 
@@ -54,6 +54,6 @@ void main() {
   publisher_test.main();
   replayer_test.main();
   subject_test.main();
-  value_subject_test.main();
+  variable_test.main();
   async_test.main();
 }
